moved transport and infrastructure classes into config folder
